feat: [CN-52] move builderpool pkg to internal/
